requirements.txt
flask~=2.0.2
python-3.7
python-telegram-bot==13.7
python-dotenv